Get rid of in-package form in globals.lisp
The long-term plan for ergolib is for it to be the core of a new
dialect of Lisp that is implemented in CL but is not backwards-compatible
with CL.  One of the things that will make ErgoLisp non-backwards-compatible
is replacing packages with lexicons.  To support the eventual transition,
the code that is intended to some day become Ergo code must be package-agnostic
for now (i.e. it must be able to run in both the cl-user package and the
(eventual) ergo package.  So it can't have an in-package form in it.